requirements.txt
django-admin
celery
django-celery
celery[redis]
python-memcached